Module zkpass_client::core

Structs

DataVerificationRequest

Represents a request that contains information needed to verify user data.

Entry

Jwk

KeysetEndpoint

Represents a JWKS (JSON Web Key Set) end point.

OutputReader

ProofMethodOutput

Contains the result of a zkPass query.

PublicKey

Represents a public key.

ZkPassProof

The value returned by the generate_zkpass_proof API of the zkPass Service.

Enums

PublicKeyOption

Provides alternative ways to represent a public key.

Val

ZkPassError

Defines various errors that come from the zkPass Service.

Traits

KeysetEndpointResolver

Defines a function for retrieving a public key from a JWKS endpoint.

Functions

decrypt_jwe_token

encrypt_data_to_jwe_token

sign_data_to_jws_token

verify_jws_token