Struct zkpass_client::interface::ZkPassClient

pub struct ZkPassClient {
    pub zkpass_service_url: String,
    pub zkpass_api_key: Option<ZkPassApiKey>,
    pub zkvm: String,
}

The main struct which implements all of the traits defined in this module.

You only need to specify the zkpass_api_key if you want to generate zkPass proofs Otherwise, you can leave them as None.

Fields
zkpass_service_url: Stringzkpass_api_key: Option<ZkPassApiKey>zkvm: String
Implementations

impl ZkPassClient

pub fn new( zkpass_service_url: &str, zkpass_api_key: Option<ZkPassApiKey>, zkvm: &str ) -> ZkPassClient

pub fn get_api_token(&self) -> Result<String, ZkPassError>

pub async fn fetch_public_keys(&self, kid: &str) -> Result<Jwk, ZkPassError>