Trait zkpass_client::interface::ZkPassProofGenerator

pub trait ZkPassProofGenerator {
    // Required method
    fn generate_zkpass_proof<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        user_data_token: &'life1 str,
        dvr_token: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<String, ZkPassError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Defines a function for generating zkPass Proof.

ZkPassProofGenerator provides an asynchronous abstraction for generating zkPass Proofs Implement this trait to define custom proof generation logic for the ZkPass service.

Implementors of this trait should handle communication with a zkPass service and perform necessary computations to generate a proof token, which can be used for secure, private authentication.

Required Methods

fn generate_zkpass_proof<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, user_data_token: &'life1 str, dvr_token: &'life2 str ) -> Pin<Box<dyn Future<Output = Result<String, ZkPassError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Description

Generates a ZkPass proof.

This function calls the corresponding RESTful API provided by the zkPass Service, where the query in the DVR is processed and the proof is generated. It encrypts both the signed user data token and the signed DVR token prior to sending it to the zkPass Service.

This function is asynchronous and must be awaited. It’s designed to be used in asynchronous contexts, for example within async functions or blocks.

Parameters

Argument	Description

user_data_token 	The signed user data JWS token

dvr_token 	The signed DVR JWS token

Return Value	Description

Result<String, ZkPassError> 	On Ok, the signed ZkPassProof token is returned as a String
On Err, the ZkPassError is returned