Trait zkpass_client::interface::ZkPassUtility

pub trait ZkPassUtility {
    // Required methods
    fn sign_data_to_jws_token(
        &self,
        signing_key: &str,
        data: Value,
        verifying_key_jwks: Option<KeysetEndpoint>
    ) -> Result<String, ZkPassError>;
    fn verify_jws_token(
        &self,
        key: &str,
        jws_token: &str
    ) -> Result<(Value, String), ZkPassError>;
    fn encrypt_data_to_jwe_token(
        &self,
        key: &str,
        data: Value
    ) -> Result<String, ZkPassError>;
    fn decrypt_jwe_token(
        &self,
        key: &str,
        jwe_token: &str
    ) -> Result<(String, String), ZkPassError>;
}

Defines functions for digital signature and encryption.

Speficially it has functions for:

signing/verifying the signature of JWT (JSON Web Signature) tokens

encrypting/decrypting the JWE (JSON Web Encryption) tokens

Required Methods

fn sign_data_to_jws_token( &self, signing_key: &str, data: Value, verifying_key_jwks: Option<KeysetEndpoint> ) -> Result<String, ZkPassError>

Description

Signs data into a JWS token.

Parameters

Argument	Description

signing_key 	The public key used to sign the data

data 	The JSON value to be signed

verifying_key_jwks 	The optional JWKS endpoint of the public key to be set in the header of the JWS token

Return Value	Description

Result<String, ZkPassError> 	On Ok, the signed JWS token is returned as a String
On Err, the ZkPassError is returned

fn verify_jws_token( &self, key: &str, jws_token: &str ) -> Result<(Value, String), ZkPassError>

Description

Verifies the signature of a JWS token.

Parameters

Argument	Description

key 	The public key used to verify the signature of the JWS token

jws_token 	The JWS token to be verified

Return Value	Description

Result<(Value, String), ZkPassError> 	On Ok, the verified value is returned as a Value, and the JWT header is also returned as a String
On Err, the ZkPassError is returned

fn encrypt_data_to_jwe_token( &self, key: &str, data: Value ) -> Result<String, ZkPassError>

Description

Encrypts data into a JWE token.

The encryption is based on ECDH-ES with GCM/AES256 encryption. The key pair used for encryption/decryption is an ECDH key pair.

Parameters

Argument	Description

key 	The public ECDH key used for the encryption purpose

data 	The JSON value to be encrypted

Return Value	Description

Result<String, ZkPassError> 	On Ok, the encrypted JWE token is returned as a String
On Err, the ZkPassError is returned

fn decrypt_jwe_token( &self, key: &str, jwe_token: &str ) -> Result<(String, String), ZkPassError>

Description

Decrypts a JWT token.

Parameters

Argument	Description

key 	The private ECDH key used to the decryption purpose

jwe_token 	The JWE token to be decrypted

Return Value	Description

Result<(String, String), ZkPassError> 	On Ok, the decrypted value is returned as a String, and the JWT header is also returned as a String
On Err, the ZkPassError is returned